src/api/routes/sundayschool.php src/EcclesiaCRM/APIControllers/SundaySchoolController.php
